+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+## [0.11.0] - 2025-11-10
+
 ### Added
 
 - Added subtraction and negation to timing
@@ -122,7 +124,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.1.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/spapanik/pyutilkit/compare/v0.10.0...main
+[Unreleased]: https://github.com/spapanik/pyutilkit/compare/v0.11.0...main
+[0.11.0]: https://github.com/spapanik/pyutilkit/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/spapanik/pyutilkit/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/spapanik/pyutilkit/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/spapanik/pyutilkit/compare/v0.7.0...v0.8.0